+++ explcheck/doc/e424.lua
@@ -1,0 +1,1 @@
+../testfiles/e424.lua