+++ explcheck/doc/w423-01.lua
@@ -1,0 +1,1 @@
+../testfiles/w423-01.lua