+++ explcheck/doc/w423-02.lua
@@ -1,0 +1,1 @@
+../testfiles/w423-02.lua